Added MainActivity and FeedActivity
