+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,6 @@
 <resources>
     <string name="app_name">MOod9</string>
     <string name="action_settings">Settings</string>
+    <string name="username">Username</string>
+    <string name="title_activity_feed">FeedActivity</string>
 </resources>
